Create main.py
Application Programming Interference provides  a URL through which we can request and get the response of the data for our further functionality in the system. It provides an access to the certain portion of data of a particular website. Request module can help in executing the url  of an API and with the help of Smtplib we can provide the notifier functionality.